Made create_user + DB Class
Made an activity for creating a user, there are some proofing, such as making sure that the numbers are at a correct lenght, but still needs some more quality control and UX update. It does however work with the DB. Also made a DB class for making it easier in the lenght.